+++ frontend/ogy_dashboard/src/hooks/metrics/useTokenDistribution.tsx
@@ -65,8 +65,6 @@ const useTokenDistribution = ({
       })
     : [];
 
-  console.log(rows);
-
   return {
     data: {
       list: {